🍿 Move More functions to common.js
