+++ api/plugins/simpleApi.js
@@ -17,38 +17,6 @@ let basePage = [];
 basePage['index.html'] = fs.readFileSync('./plugins/simpleApi/index.html');
 basePage['index.css'] = fs.readFileSync('./plugins/simpleApi/index.css');
 
-// Get Request
-function get(uri) {
-    return new Promise((resolve, reject) => {
-        let req = http.get(uri, response => {
-            let data = '';
-            response.on('data', chunk => (data += chunk));
-            response.on('end', () => resolve(data));
-        });
-        req.on('error', reject);
-    });
-}
-
-let cache = {};
-
-function getData(url) {
-    return new Promise((resolve, reject) => {
-        if (
-            url in cache &&
-            Date.now() - cache[url].time <= pluginConfig.cacheTime
-        )
-            resolve([cache[url].data, true]);
-        else {
-            get(url)
-                .then(data => {
-                    cache[url] = { data: data, time: Date.now() };
-                    resolve([data, false]);
-                })
-                .catch(reject);
-        }
-    });
-}
-
 function api(app, wsServer, config) {
     // Api Docs
     app.get('/api', (req, res) => {
@@ -69,15 +37,15 @@ function api(app, wsServer, config) {
     // Get current temperature
     app.get('/api/temp', (req, res) => {
         common.log('🌐 GET: /api/temp', '', req.ip);
-        getData(`http://${pluginConfig.ip}:${pluginConfig.port}/temp`)
+        common.getData(`http://${pluginConfig.ip}:${pluginConfig.port}/temp`)
             .then(d => {
                 data = JSON.parse(d[0]);
                 res.send({ temp: data.temp, cached: d[1] });
             })
-            .catch(err => {
-                res.status(500);
-                res.send({ error: err });
-            });
+            // .catch(err => {
+            //     res.status(500);
+            //     res.send({ error: err });
+            // });
     });
 
     // Get the temperature at a specific time
@@ -85,7 +53,7 @@ function api(app, wsServer, config) {
         common.log('🌐 GET: /api/temp/:time', '', req.ip);
         let time = new Date(parseInt(req.params.time) * 1000);
         let temp = null;
-        getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/download`)
+        common.getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/download`)
             .then(d => {
                 data = {};
                 d[0].split('\n').forEach(e => {
@@ -110,7 +78,7 @@ function api(app, wsServer, config) {
     // Get temperature stats
     app.get('/api/stats', (req, res) => {
         common.log('🌐 GET: /api/stats', '', req.ip);
-        getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/stats`)
+        common.getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/stats`)
             .then(d => {
                 data = JSON.parse(d[0]);
                 res.send({
@@ -131,7 +99,7 @@ function api(app, wsServer, config) {
     // Get temperature history
     app.get('/api/history', (req, res) => {
         common.log('🌐 GET: /api/history', '', req.ip);
-        getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/download`)
+        common.getData(`http://${pluginConfig.ip}:${pluginConfig.port}/data/download`)
             .then(d => {
                 data = {};
                 let lines = d[0].split('\n');
